turn demo into full fledged lib
