+++ Cargo.toml
@@ -1,14 +1,24 @@
 [package]
 name = "tokio-socks5"
 version = "0.1.0"
-authors = ["Alex Crichton <alex@alexcrichton.com>"]
+authors = ["Alex Crichton <alex@alexcrichton.com>",
+           "heartsucker <heartsucker@autistici.org>"]
+
+[lib]
+name = "tokio_socks5"
+path = "src/lib.rs"
+
+[[bin]]
+name = "tokio-socks5"
+path = "bin/tokio_socks5.rs"
 
 [dependencies]
-log = "0.3"
+clap = "2"
+env_logger = "0.5"
 futures = "0.1"
+log = "0.4"
 tokio-core = "0.1"
 tokio-io = "0.1"
-env_logger = "0.3"
 trust-dns = { version = "0.9.2", default-features = false }
 
 [dev-dependencies]
