+++ src/lib.rs
@@ -1,0 +1,9 @@
+extern crate futures;
+#[macro_use]
+extern crate log;
+#[macro_use]
+extern crate tokio_core;
+extern crate tokio_io;
+extern crate trust_dns;
+
+pub mod server;
